requirements/dev.txt
-r base.txt
django-debug-toolbar==1.2.1
# Required by django-debug-toolbar
sqlparse==0.1.11

coverage==3.7.1
pylint==1.1.0
# Required by pylint
logilab-common==0.61.0
astroid==1.0.1

flake8==2.1.0
  mccabe==0.2.1
  pep8==1.5.6
  pyflakes==0.8.1

factory_boy==2.3.1
Mock==1.0.1
